+++ modules/extensions-spring-support/src/test/resources/validation/dynamic-stateful-override-config.xml
